Move towards using ZDCF format instead of my own
Duh, I wish I had found that earlier...

Anyway, this commit ditches the old JSON-based file format and adopts
ZDCF, albeit incompletely. ZDCF does not specify an encoding. It just
defines the data structure and its contents in terms of hashes and
arrays and basic types. Thus, we abstract away the encoding part and
ship JSON by default.

Furthermore, this introduces a ZDCF validator based on Rx. This should
make it rather easy to write the code that builds the actual objects
from the definitions.

What remains to be decided (and then implemented) is whether validated
ZDCF is "good enough" as an in-memory data structure for non-live
representation or whether the "Device", "Socket", etc. "Declare"-style
classes are still useful.